refactor of specimen

+++ backend/camellia/src/main/java/com/camellia/services/requests/CultivarRequestService.java
@@ -24,6 +24,6 @@ public class CultivarRequestService {
     }
 
     public CultivarRequest getCultivarRequestById(long id) {
-        return repository.findById((long) id);
+        return repository.findById(id);
     }
 }
